+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - A fallback value parsing from the \<paint\> type.
 - Parse the `inherit` value from the `font-family` attribute.
 
+### Changed
+- `path::Segment` now returns only command char and segment data. All segment manipulation methods are removed.
+
+### Removed
+- `path::Command` struct.
+
 ## [0.0.2] - 2016-09-09
 ### Removed
 - Remove `BlockProgression` and `TextAlign` from `AttributeId` enum since they are not a SVG attributes.